backend/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
PyJWT
flask-cors